Unmarshallers didn't allow a complex sub-attribute that is being used by the resource schema. This caused the resource schema downloaded from the server not having any sub-attributes defined.
The HttpUrlConnection used by default by Wink client behaves badly when connections are opened and closed quickly and repeatedly. It relies on the finalizer to close the socket, which can lag behind. A new constructor for SCIMServices that allows the user to configure the Wink client themselves. In the query rate tool, it now uses the Apache HttpClient library instead for Wink. It will now also pool connections up to the number of threads.

JiraIssue: SCIM-78
JiraIssue: SCIM-79