===BEGIN-RELEASE-NOTE=== Update the SCIMEndpoint to catch RuntimeExceptions and wrap them with a SCIMException. ===END-RELEASE-NOTE===
===BEGIN-INTERNAL-COMMENT===
There are also some minor POM dependency refactorings, mainly to make the SCIM-SDK depend on the wink-client-apache-httpclient module. Most consumers will want to use the Apache HTTPClient library and our examples in the Getting Started Guide use it as well. This library transitively pulls in the wink-client and Apache httpclient libraries.
===END-INTERNAL-COMMENT===

JiraIssue: SCIM-236