===BEGIN-RELEASE-NOTE=== Comply with the for SCIM 1.1 schema changes: - phoneNumbers attribute now uses the RFC3966 syntax. - The Group displayName attribute now required. - The Group members attribute is no longer required.
To accommodate for possible differences in LDAP's phone number format, several arguments are added to the new TelephoneNumberTransformation so that the mapping behavior may be customized to fit installation environment. Details about the arguments are in the resources.xml file. 
===END-RELEASE-NOTE===

To format phone numbers using RFC3966, I added a new RFC3966Transformation implementation using Google's libphonenumber library: 

http://code.google.com/p/libphonenumber/ 

This makes the parsing very robust and handles most varieties of numbers from different countries and languages. The transformation tries its best to canonicalize the values but will fall back to just passing the string as is if all else fails.

The last two issues are just simple updates to the CoreSchema file and the default resources.xml.

Reviewer: tthompson
Product: scim
JiraIssue: DS-6892