_Problem_ Using a filter that includes a simpleMultiValued attribute from a custom schema will return no resource data even when the ldap filter matches.
_Analysis_
simpleMultiValued attributes have an implied set of sub-attributes labeled "value", "type", etc.  These sub attrs are added automatically when the attribute descriptor is constructed.  These sub attrs are created as "core" schema attributes.  When a filtered request is serviced from LDAP and we map the result to a scim resource object we check the original filter to ensure that the results match the new resource.  Since the simpleMultiValued attribute is associated with a custom schema then the filter expects the new resource (result entry) to match that schema, however the sub attr "value" in this situation (which we are matching against) is part of the core schema, so the matching fails.  Thus no resource data is returned.

_Correction_
The constructor for simpleMultiValued objects calls a utility method on the Core Schema that adds the implied sub attributes like "value", "type", etc.  I've changed this method to take a schema argument, which is used in place of the hard coded "core" schema value that was being applied to each of the sub attrs.  Now the same schema used to construct the simpleMultiValued attr will be applied to the sub attrs.

_Test_
I was able to reproduce this issue consistently, once I determined that you need to use a custom schema with a simpleMultiValued attr.  "emails" from the core schema is a easy example to copy.  Using the core "Users" endpoint with a filter for "emails.value eq ..." the results would be returned as expected, and using my custom endpoint "ExUsers" would return a empty data object but still note the matching result.
I experimented with a couple of unit tests for this issue, which probably requires a more end to end test, but the primary hurdle was the way that we have centered the tests around the CoreSchema class and the way it statically creates the objects we use for testing.  I've opened DS-8227 to possibly enhance this.  Seeing as how the manual testing is fairly straight forward, I feel like my testing there is fairly indicative that this issue has been resolved.

===BEGIN-RELEASE-NOTE===
Fix a bug in SCIM that prevented simpleMultiValued attributes from returning matching resources when using a custom schema urn.
===END-RELEASE-NOTE===

Reviewer: Andy Coulbeck

JiraIssue:DS-8226

SalesforceCaseNumber:1883

Product:ds