Refactor the versioning support code in the SCIM-SDK to support If-Match and If-None-Match headers for GET, PUT, PATCH, and DELETE operations.
Correctly handle the case where the modify timestamp is not present in the LDAP entry by using "0" as the etag.

Remove the requirement to have the meta.version attribute for PUT and PATCH operations in a Bulk request. 

JiraIssue: DS-10091
JiraIssue: DS-10090
JiraIssue: DS-10080
JiraIssue: DS-10078
Product: scim-sdk
Reviewer: coulbeck