Updated JsonUnmarshaller.unmarshalResources(...) to handle a non-list response
Updated JsonUnmarshaller.unmarshalResources(...) so that it will make a
best-effort attempt to unmarshal a JSON entity to a Resources object, even if
the JSON entity is not formatted as a SCIM list response but is instead
formatted as a singular resource. This addresses a potential interoperability
issue in which a SCIM client calls SCIMEndpoint.query(...) for a SCIM server's
"/Schemas" endpoint, assuming that it will receive a list response, but the
SCIM server instead returns a single schema resource.

JiraIssue: DS-36326