Make JSON default response content type
According to RFC7644, when no format is provided
by accept header in the request, the server must
default to JSON response format. A few of our
endpoints do not currently do this, so making
this the default for those endpoints.

===BEGIN-RELEASE-NOTE===
Made JSON the default SCIM response content type
for requests that did not specify an accept content
type.
===END-RELEASE-NOTE===

Reviewer: Jacob Childress

JiraIssue: DS-40869

SalesforceCaseNumber: 00672865

Product: ds